feat : add products

+++ src/components/ui/CartCounter.tsx
@@ -4,6 +4,8 @@ import React, { useState } from "react";
 import { Button } from "../../components/ui/button";
 import { FaMinus, FaPlus } from "react-icons/fa6";
 import { cn } from "../../lib/utils";
+import { useAppDispatch } from "../../lib/hooks/redux";
+import { updateStock } from "@/lib/features/products/productsSlice"; // นำเข้า updateStock จาก productsSlice
 
 type CartCounterProps = {
   isZeroDelete?: boolean;
@@ -11,6 +13,7 @@ type CartCounterProps = {
   onRemove?: (value: number) => void;
   className?: string;
   initialValue?: number;
+  productId: number; // เพิ่ม productId เพื่อลด stock
 };
 
 const CartCounter = ({
@@ -19,24 +22,23 @@ const CartCounter = ({
   onRemove,
   className,
   initialValue = 1,
+  productId,
 }: CartCounterProps) => {
   const [counter, setCounter] = useState<number>(initialValue);
+  const dispatch = useAppDispatch();
 
   const addToCart = () => {
-    if (onAdd) {
-      onAdd(counter + 1);
-    }
-    setCounter(counter + 1);
+    setCounter((prev) => prev + 1);
+    dispatch(updateStock({ id: productId, change: 1 })); // ลด stock ลง 1
+    if (onAdd) onAdd(counter + 1);
   };
 
   const remove = () => {
     if ((counter === 1 && !isZeroDelete) || counter <= 0) return;
 
-    if (onRemove) {
-      onRemove(counter - 1);
-    }
-    if (counter - 1 <= 0) return;
-    setCounter(counter - 1);
+    setCounter((prev) => prev - 1);
+    dispatch(updateStock({ id: productId, change: -1 })); // เพิ่ม stock คืน 1
+    if (onRemove) onRemove(counter - 1);
   };
 
   return (
@@ -51,19 +53,17 @@ const CartCounter = ({
         size="icon"
         type="button"
         className="h-5 w-5 sm:h-6 sm:w-6 text-xl hover:bg-transparent"
-        onClick={() => remove()}
+        onClick={remove}
       >
         <FaMinus />
       </Button>
-      <span className="font-medium text-sm sm:text-base">
-        {!isZeroDelete ? counter : initialValue}
-      </span>
+      <span className="font-medium text-sm sm:text-base">{counter}</span>
       <Button
         variant="ghost"
         size="icon"
         type="button"
         className="h-5 w-5 sm:h-6 sm:w-6 text-xl hover:bg-transparent"
-        onClick={() => addToCart()}
+        onClick={addToCart}
       >
         <FaPlus />
       </Button>
